python_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
litellm==1.41.0
anthropic==0.34.0
openai==1.35.0
google-generativeai==0.7.0
pydantic==2.5.0
python-dotenv==1.0.0
